requirements.txt
sentence-transformers==2.2.2
faiss-cpu==1.7.4
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
fastapi-limiter==0.1.5
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.3
httpx==0.25.1 